ajout affichage main

+++ td2/src/fonction/utile/Tangeante.java
@@ -6,9 +6,9 @@ public class Tangeante implements Fonction1D {
 
     Fonction1D f1d;
     Fonction1D derivee;
-    int x0;
+    double x0;
 
-    public Tangeante(Fonction1D f1d, int x0){
+    public Tangeante(Fonction1D f1d, double x0){
         this.f1d = f1d;
         this.derivee = new Derivee(f1d, Derivee.DX);
         this.x0 = x0;
